Fixed another early stopping bug
-Hopefully the last. This one was caused by using the on_validation_end
hook instead of the on_train_epoch_end hook. Issue is that this former
hook gets called prior to the start of training which means that
the "best_score" will potentially be a value that's not even seen
in the wandb graphs. This resulted in early stopping happening
after 5 epochs for many of the runs.